# eidStack-CMU/openapi/openapi.yaml
openapi: 3.0.0
paths:
...
  /issuance/offerStatus:
    get:
      operationId: IssuanceController_getOfferStatus
      parameters:
        - name: credentialExchangeId
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Proof state (string)
        '400':
          description: Invalid credentialExchangeId
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      summary: Get proof/request status
      tags: &ref_2
        - issuance
  /issuance/listSchemas:
    get:
      operationId: IssuanceController_listSchemas
      parameters: []
      responses:
        '200':
          description: Schemas returned successfully
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      summary: List all schemas (non-paginated)
      tags: *ref_2
  /issuance/schemas:
    get:
      operationId: IssuanceController_listSchemasPaginated
      parameters:
        - name: limit
          required: false
          in: query
          schema:
            example: 10
            type: number
        - name: page
          required: false
          in: query
          schema:
            example: 1
            type: number
      responses:
        '200':
          description: Paginated list of schemas
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      summary: List schemas (paginated)
      tags: *ref_2
    post:
      operationId: IssuanceController_createSchema
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSchemaInput'
      responses:
        '201':
          description: Schema created successfully
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      summary: Create a new schema
      tags: *ref_2
  /issuance/credential-definitions:
    get:
      operationId: IssuanceController_listCredentialDefinitionsPaginated
      parameters:
        - name: limit
          required: false
          in: query
          schema:
            example: 10
            type: number
        - name: page
          required: false
          in: query
          schema:
            example: 1
            type: number
      responses:
        '200':
          description: Paginated list of credential definitions
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      summary: List credential definitions (paginated)
      tags: *ref_2
    post:
      operationId: IssuanceController_createCredentialDefinition
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCredDefInput'
      responses:
        '201':
          description: Credential definition created successfully
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      summary: Create a new credential definition
      tags: *ref_2
  /issuance/offer:
    post:
      operationId: IssuanceController_offerCredentialWithUrl
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OfferCredentialWithUrlDto'
      responses:
        '201':
          description: Credential offer created successfully
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      summary: Offer credential (Out-of-Band)
      tags: *ref_2
...
info:
  title: e-ID REST API
  description: REST APIs converted from GraphQL
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    AttributeInput:
      type: object
      properties:
        attributeName:
          type: string
        schemaDataType:
          type: string
        displayName:
          type: string
      required:
        - attributeName
        - schemaDataType
        - displayName
    CreateSchemaInput:
      type: object
      properties:
        name:
          type: string
        version:
          type: string
        attributes:
          type: array
          items:
            $ref: '#/components/schemas/AttributeInput'
        issuerId:
          type: string
        orgId:
          type: string
      required:
        - name
        - version
        - attributes
    CreateCredDefInput:
      type: object
      properties:
        schemaId:
          type: string
        tag:
          type: string
        supportRevocation:
          type: boolean
      required:
        - schemaId
        - tag
    CredentialAttributeInput:
      type: object
      properties:
        name:
          type: string
        value:
          type: string
      required:
        - name
        - value
    OfferCredentialWithUrlDto:
      type: object
      properties:
        credentialDefinitionId:
          type: string
        connectionId:
          type: string
        attributes:
          type: array
          items:
            $ref: '#/components/schemas/CredentialAttributeInput'
        comment:
          type: string
      required:
        - credentialDefinitionId
...
    ErrorResponse:
      type: object
      properties:
        correlationId:
          type: string
        statusCode:
          type: integer
          format: int32
        error:
          type: string
        message:
          oneOf:
            - type: string
            - type: array
            - type: object